feat(ui): ship the design gates as runnable package scripts
Move the contrast and no-raw-hex gates into packages/ui/scripts and
expose them as package bins (repowise-ui-gates, repowise-ui-no-raw-hex,
repowise-ui-contrast-check) so consumers run the same canonical checks
from node_modules instead of forking them. Both gates are extensible:
no-raw-hex takes --root and repeatable --allow entries; contrast takes
repeatable --extend JSON files adding consumer tokens and pairs. The
docs/design entry points remain as thin wrappers with identical default
behavior.
