+++ packages/ui/package.json
@@ -7,7 +7,8 @@
   "types": "./src/index.ts",
   "files": [
     "src",
-    "styles"
+    "styles",
+    "scripts"
   ],
   "sideEffects": [
     "./styles/globals.css",
@@ -90,7 +91,8 @@
   },
   "scripts": {
     "type-check": "tsc --noEmit",
-    "test": "vitest run"
+    "test": "vitest run",
+    "gates": "bash ./scripts/ui-gates.sh"
   },
   "dependencies": {
     "@lottiefiles/dotlottie-react": "^0.19.4",
@@ -147,5 +149,10 @@
     "react-dom": "^19.0.0",
     "typescript": "^5.6.0",
     "vitest": "^4.1.5"
+  },
+  "bin": {
+    "repowise-ui-gates": "./scripts/ui-gates.sh",
+    "repowise-ui-no-raw-hex": "./scripts/no-raw-hex-check.sh",
+    "repowise-ui-contrast-check": "./scripts/contrast-check.py"
   }
 }
